More idiomatic source layout
This is an attempt to make the source layout adhere to Go conventions,
for example outlined here:
  http://code.google.com/p/go-wiki/wiki/GithubCodeLayout

Basically, we can assume that people will name their packages using
`github.com/username/...` pattern and put the source files in the
root of repository. It is not desirable to force them to use
non-standard directory layout or change their `GOPATH` locally
to get the Shippable build to work.

To make the standard layout work on Shippable,  we need to link the
checked-out repository to `$GOPATH/src`, so it will be properly resolved
by Go tools. When testing the main package, we can use `./` as the package
name to avoid repetition. If we had multiple package tests, we could use
`./...`.

BTW, I couldn't get `gocover-cobertura` nor `gocov-xml` to output
anything more interesting than '0' as line or branch coverage. I guess
this is a problem with current versions of these tools.

Also, modified slightly the test suite file, so the path to the test
results location is taken from an environment variable.